% Auto-generated: do not edit by hand
\name{''Modelviewer}

\alias{''Modelviewer}

\title{Modelviewer component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''Modelviewer(id=NULL, alt=NULL, exposure=NULL, src=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{alt}{Character. The src displayed in the input.}

\item{exposure}{Character. The src displayed in the input.}

\item{src}{Character. The src displayed in the input.}

\item{style}{Named list. The src displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

