% Auto-generated: do not edit by hand
\name{''ModelViewer}

\alias{''ModelViewer}

\title{ModelViewer component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''ModelViewer(id=NULL, alt=NULL, exposure=NULL, label=NULL, src=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{alt}{Character. The src displayed in the input.}

\item{exposure}{Character. The src displayed in the input.}

\item{label}{Character. A label that will be printed when this component is rendered.}

\item{src}{Character. The src displayed in the input.}

\item{style}{Named list. The src displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

